Move FDK parsing into FDKData.js
